Previous/Next experiment and Second Scene experiment
This commit demonstrates two features I would like to develop further.
First, I have added two Labels to the form, each with an onClick event
These labels would be used to navigate the NHL App to a previous or next date.
All I wanted to see here was what the labels would look like and could I attach event handlers to labels.
This was a successful experiment.

The second experiment was to see whether I could add a second scene to the Stage
I saved the Stage as an object property
I changed the event for one of the old buttons to call a new method.
This new method "recreateStage" defined a new set of controls, a new pane, and a new scene
Then it set the original stage to the new scene
This was also successful

These two tools can be used to create individual scenes for each date and scroll between them easily

This also suggests that I should have a single-purpose method that develops the scene for any given date.